### 2023-04-24 #5282 @tenemark
Hello, sorry about the delay in my response. I have been travelling. MIKE SHE is indeed a proprietary application. I will have to ask around to see if I can provide a license. I will get back to you in a couple of days.
