+++ mikecalsetup/ost_post_proc.py
@@ -1,0 +1,337 @@
+# -*- coding: utf-8 -*-
+"""
+OSTRICH output reading
+Example script reading all solutions from an ongoing or finished OSTRICH
+ParaPADDS calibration
+
+Raphael Schneider, rs@geus.dk, GEUS Hydro, Mar 2023
+"""
+
+import os
+import pandas as pd
+import numpy as np
+import mikecalsetup.ost_file as ost_file
+import glob
+import seaborn as sns
+import shutil
+
+
+class OstPostProc:
+    """Postprocessing of OSTRICH files for ParaPADDS calibration."""
+
+    def __init__(self, root):
+        self.root = root
+        self.out_fp = root + r'OstOutput0.txt'
+        self.outf_fp = root + r'OstModel'
+        self.oin_fp = root + r'ostIn.txt'
+
+        # extract output data
+        self.ns = self.read_ostoutput0()
+        self.fs = self.read_ostmodel_files()
+        self.ws = self.get_observation_weigths()
+
+    def read_ostoutput0(self):
+        """
+        Read ostoutput file and records non-dominant solutions in dataframe.
+
+        Returns
+        -------
+        ns : pandas dataframe
+            Non-dominant solutions from ostoutputfile
+
+        """
+        with open(self.out_fp, 'r') as f:
+            lines = f.readlines()
+
+        # done = 1 if the run is finished
+        temp = [i for i, line in enumerate(lines)
+                if 'Non-Dominated Solutions\n' == line]
+        done = 1 if len(temp) != 0 else 0
+
+        # loading last set of non-dominated solutions
+        if done == 0:
+            # starting index last set of non-dominated solutions
+            lino = [i for i, line in enumerate(lines)
+                    if '\n' == line][-1]
+            # column names
+            idx = [i for i, line in enumerate(lines)
+                   if 'Ostrich Run Record\n' == line][0]
+            cols = lines[idx+1].split()
+            cols = [c for c in cols if (c.find('(') > -1) | (c.find('__') > -1)]
+
+            # loading data
+            ns = pd.read_csv(self.out_fp, skiprows=lino+1, header=None, sep='\s+')
+            ns = ns[ns.columns.tolist()[1:-1]]
+            ns.columns = cols
+        elif done == 1:
+            # starting index
+            idx = temp[0]
+            last_idx = [idx+i for i, line in enumerate(lines[idx:])
+                        if '\n' == line][0]
+
+            # columns
+            cols = lines[idx+1].split()
+            cols = [col for col in cols if ')' != col]
+            cols = [c for c in cols if (c.find('(') > -1) | (c.find('__') > -1)]
+
+            # loading data
+            ns = pd.read_csv(self.out_fp, skiprows=temp[0]+2, header=None, sep='\s+',
+                             skipfooter=len(lines)-last_idx)
+            ns.columns = cols
+        return ns
+
+    def read_ostmodel_files(self):
+        """
+        Record everything from ostmodel files, i.e. full set of solutions.
+
+        Includes single observations which are not present in ostoutput file.
+
+        Returns
+        -------
+        fs : pandas dataframe
+            all records from ostmodel files
+
+        """
+        fs = pd.DataFrame()
+        for file in glob.glob(self.root + "OstModel*"):
+            fs = pd.concat([fs, pd.read_csv(file, sep='\s+',
+                                            index_col=0, engine='python')])
+        fs.reset_index(inplace=True)
+        return fs
+
+    def get_observation_weigths(self):
+        """
+        Get observation weigths.
+
+        Returns
+        -------
+        ws : TYPE
+            DESCRIPTION.
+
+        """
+        # read ostin file
[…]
