### 2023-05-03 #5282 @kthyng
@tenemark Hi, I am another editor here at JOSS. It would help a lot if you could dig up some reviewers from your extended network who are not conflicted with you to suggest. Based on past experience, it's very difficult to find reviewers for proprietary software and for one specific operating system, particularly if this isn't the exact field of any of our editors. If we cannot find reviewers, we will not be able to review your submission, unfortunately. Just wanted to set expectations reasonably for how this might go.
