### 2023-06-03 #5282 @jsta
:wave: @Jsmariegaard Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

There are a few details you should know about this particular submission namely that there is a dependency on external software that only runs on Windows and necessary configuration may be complex.
