### 2023-06-08 #5282 @ecomodeller
Hi @jsta, I have access to a MIKE SHE license.

What do I do next?
