### 2023-06-09 #5282 @jsta
@ecomodeller That's great to hear, thanks for volunteering. You can start by entering the command `@editorialbot generate my checklist` and checking the boxes that meet the specified criteria. https://joss.readthedocs.io/en/latest/review_criteria.html will give more details. Please let me know if you have questions.
