### 2023-06-12 #5548 @editorialbot
[REVIEW]: mikecalsetup: an open-source python tool for automatically creating calibration setups of MIKE SHE models in PEST or OSTRICH
**Submitting author:** @tenemark (<a href="http://orcid.org/0000-0002-3881-2739">Trine Enemark</a>)
**Repository:** https://github.com/tenemark/mikecalsetup
**Branch with paper.md** (empty if default branch):
**Version:** v0.1
