### 2023-06-15 tenemark/mikecalsetup#1 @aleaf
Documentation improvement: package requirements
Hi, thanks for the opportunity to [review this software package](openjournals/joss-reviews#5548). Installing the package as directed appeared to work OK, and I can see in setup.py that [mikeio](https://github.com/DHI/mikeio) is a dependency. However mikeio in turn depends on a number of other packages, which themselves have extensive library dependencies.

It would be helpful to users to make the dependencies more explicit. Often this is done within a [pyproject.toml file](https://github.com/DHI/mikeio/blob/main/pyproject.toml), [requirements.txt](https://github.com/NSLS-II/scientific-python-cookiecutter/blob/main/requirements.txt), or [conda environment file](https://github.com/geopandas/geopandas/blob/main/environment.yml). At a minimum, numpy, pandas and mikeio should be specified, (and any other non-standard library packages that are imported directly), but it would also be good to also include any dependencies of mikeio that the functionality of this package depends on. A conda environment file would be ideal, as it would provide another robust option for setting up the required environment if there are issues with the pip install. In the past, pip hasn't worked well for installing packages with external library dependencies on Windows, especially geospatial packages such as geopandas and rasterio. Note that conda environment files can include a pip installs section, so packages such as numpy, pandas, geopandas, etc could be installed by conda (or [mamba](https://mamba.readthedocs.io/en/latest/)), and mikeio could be installed last via pip.
