### 2023-06-15 tenemark/mikecalsetup#2 @aleaf
example usage.ipynb requires jupyter and notebook to be installed
these could be included in an environment.yml file as mentioned in #1.
