### 2023-06-15 tenemark/mikecalsetup#3 @aleaf
test example yields OSError
After installing mikecalsetup as directed in the readme, executing the tests from the tests folder with:
```
python -m unittest
```
produced this error:
```
OSError: [Errno 22] Invalid argument: 'example_models\\Karup\\Karup_Basic\\|.\\Model Inputs\\MIKE_SHE_vege.ETV|'
```

I also encountered this error when trying to run `example usage.ipynb` after installing `jupyter` and `notebook` as mentioned in #2.
