### 2023-06-15 tenemark/mikecalsetup#4 @aleaf
Documentation improvement: statement of need
From the [JOSS checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation):

**A statement of need:** The existing readme file starts with this, but I think it would be helpful to flesh it out some more.
* How is setting up parameter estimation for a MIKE SHE model cumbersome? Are the input and output formats non-standard (i.e. not simple text arrays or tabular data that could be handled with [the PstFrom module in pyEMU](https://github.com/pypest/pyemu/blob/develop/examples/pstfrom_mf6.ipynb)? (Note that while PstFrom was admittedly developed around MODFLOW, more recent generalization should allow it to work with any model, though some intermediate processing may be needed).
* more generally, what kind of parameterization does mikecalsetup implement? It looks like there is some discussion of this in the example notebook, but this should also be summarized up front. Does it do spatially distributed parameterization (zones, pilot points, etc) or are the parameters tied to specific units within the MIKE SHE input structure (like a hydrologic response unit in PRMS). Apologies for knowing nothing about how MIKE SHE works. Perhaps the potential userbase (of MIKE SHE users) will have an easier time understanding the problem that is being solved here.
* it looks like some of the scope of this package is described in supported_settings.md; this file could be referenced in the documentation to better describe what mikecalsetup does, and what it's limitations are.
* similarly, how are observations handled? Is there any processing done to align them with real-world observations, or do the real-world observations need to be processed independently to align with the model output? For example annual and monthly means, baseflow separation, head and flow differences, etc.
