### 2023-06-15 tenemark/mikecalsetup#5 @aleaf
Documentation improvement: PEST example
The main readme file and JOSS paper state that mikecalsetup can produce input for PEST but `example usage.ipynb` only seems to illustrate and OSTRICH use case. It would be helpful to include a PEST example as well; this could also help users understand what the key differences are between mikecalsetup and [pyEMU](https://github.com/pypest/pyemu). Ideally this PEST example (and the existing OSTRICH example) would be complete working examples that include successful execution of a forward model run, to demonstrate that the input produced by mikecalsetup works as intended. These examples would also ideally be included in the automated testing and CI #7.
