### 2023-06-15 tenemark/mikecalsetup#7 @aleaf
Testing improvements: pytest, CI and test coverage
While the project appears to include some unit tests, I would highly encourage you to adopt pytest, which as I understand is is much more powerful and has become the defacto standard in the scientific python community.

I would also strongly encourage setup of continuous integration testing and test coverage reporting, so that the test suite is run each time a commit is pushed or a pull request is made. Test coverage reporting can help everyone better understand the strengths and weaknesses of the software, as well as the potential implications of any changes made. CI testing can save you and users time by preventing bugs and can also help others contribute to the software (and for you to accept those contributions) with confidence.

The [Scientific Python Cookiecutter](https://github.com/NSLS-II/scientific-python-cookiecutter) has an excellent template and guide to help authors set up pytest and CI with test coverage reporting relatively quickly. Regrettably, for reasons I don't understand, the CI portion of the Scientific Python Cookiecutter guide are still focused on Travis, which has largely been dropped as the standard for open-source projects in favor of GitHub Actions. But, GitHub Actions can generally be readily set up for automated testing by copying and pasting workflows from other projects (for example, you could copy the workflows from mikeio).
