### 2023-06-15 tenemark/mikecalsetup#8 @aleaf
Documentation improvement: community guidelines
The documentation could use some community guidelines, as described in the [JOSS checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#documentation). These could be included in the HTML documentation (#6).
