### 2023-06-15 #5548 @aleaf
The readme has a statement of need, but it could be improved: tenemark/mikecalsetup#4
