### 2023-06-15 #5548 @aleaf
The package needs a list of dependencies, ideally in the form of a conda environment file: tenemark/mikecalsetup#1
