### 2023-06-15 tenemark/mikecalsetup#7 @aleaf
ideally, the automated testing and CI would include a full working example that includes execution of a forward model with PEST and OSTRICH #5.
