### 2023-06-15 #5548 @aleaf
While the package includes a minimal working example, it could ideally be improved to be a complete working example that included execution of a forward model run with PEST or OSTRICH tenemark/mikecalsetup/#5.
