### 2023-06-15 #5548 @aleaf
Most functions and classes are documented with numpy-style docstrings, but I would highly encourage the author to add HTML documentation, which can be easily set up with the [Scientific Python Cookiecutter](https://github.com/NSLS-II/scientific-python-cookiecutter) or perhaps other tools tenemark/mikecalsetup#7.
