### 2023-06-15 #5548 @aleaf
While there is one test file, it didn't execute successfully, and there is no test coverage reporting or CI tenemark/mikecalsetup#3, tenemark/mikecalsetup#7.
