### 2023-06-15 #5548 @aleaf
The documentation does not have community guidelines tenemark/mikecalsetup#8
