### 2023-06-15 #5548 @aleaf
It looks like there are some issues with the reference DOIs, as flagged by @editorialbot. pyEMU should probably be cited:

White, J.T., Fienen, M.N., Doherty, J.E., 2016. A python framework for environmental model uncertainty analysis. Environmental modeling and Software 85, 217–228. https://doi.org/10.1016/j.envsoft.2016.08.017.
