### 2023-07-24 #5548 @tenemark
Hi, I understand the comments and will work on improving the package and
manuscript accordingly. Thank you for the helpful suggestions. I will get
back to you in a couple of weeks.

On Sat, Jul 8, 2023 at 4:00 AM Jemma Stachelek ***@***.***>
wrote:

> Thank you to both reviewers for their careful and thorough work!
>
> @tenemark <https://github.com/tenemark> please let myself or the
> reviewers know if you need clarification on any of the points they've
> raised.
>
