### 2023-08-11 c379aaa
Revision and extra examples
Following review in joss
