+++ joss/paper.bib
@@ -1,49 +1,3 @@
-@techreport{Matott2017,
-author = {Matott, L. Shawn},
-institution = {University at Buffalo Center for Computational Research},
-pages = {1--73},
-title = {{OSTRICH: an Optimization Software Tool, Documentation and User's Guide, Version 17.12.19}},
-url = {http://www.civil.uwaterloo.ca/envmodelling/Ostrich.html},
-year = {2017}
-}
-@techreport{Harbaugh2017,
-author = {Harbaugh, A.W. and Langevin, C.D. and Hughes, J.D. and Niswonger, R.N. and Konikow, L. F.},
-doi = {http://dx.doi.org/10.5066/F7RF5S7G},
-institution = {USGS},
-title = {{MODFLOW-2005 version 1.12.00, the U.S. Geological Survey modular groundwater model: U.S. Geological Survey Software Release}},
-year = {2017}
-}
-@techreport{Doherty2015,
-address = {Brisbane, Australia},
-author = {Doherty, J. E.},
-institution = {Watermark Numerical Computing},
-pages = {1--236},
-title = {{Calibration and uncertainty analysis for complex environmental models}},
-year = {2015}
-}
-@article{Abbott1986,
-author = {Abbott, M.B. and Bathurst, J.C. and Cunge, J.A. and O'Connell, P.E. and Rasmussen, J.},
-file = {:C\:/Users/wqg436/OneDrive - K{\o}benhavns Universitet/Papers/Abott et al. 1986.pdf:pdf},
-issn = {00221694},
-journal = {Journal of Hydrology},
-pages = {61--77},
-title = {{An introduction to the European hydrological system – systeme hydrologique Europeen, She. 2. Structure of a physically-based, distributed modeling system}},
-volume = {87},
-year = {1986}
-}
-@article{WHITE2021105022,
-abstract = {Computer models of environmental systems routinely inform decision making for water resource management. In this context, quantifying uncertainty in the important simulated outputs, and reducing uncertainty through assimilating historic system-state observations, is as important as the numerical model. However, implementing high-dimensional and stochastic workflows are challenging, often requiring that practitioners have theoretical and practical understanding of several advanced topics. Worse, implementing these important analyses can take substantial time and effort. This additional effort is often cited as justification for postponing, or even forgoing, these analyses. Herein, we present scripting tools to facilitate the efficient and repeatable construction of high-dimensional, geostatistical-based PEST interfaces, including uncertainty analyses. As demonstrated, these tools can be applied with minimal effort to a model with varied temporal and spatial discretization. Ultimately, these tools can enable low-cost access to valuable decision-support analyses earlier and more frequently during the modeling workflow.},
-author = {White, Jeremy T and Hemmings, Brioch and Fienen, Michael N and Knowling, Matthew J},
-doi = {https://doi.org/10.1016/j.envsoft.2021.105022},
-issn = {1364-8152},
-journal = {Environmental Modelling & Software},
-keywords = { Automating workflows, Cognitive load, Decision support, Model criticism, Prior-data conflict,Uncertainty analysis},
-pages = {105022},
-title = {{Towards improved environmental modeling outcomes: Enabling low-cost access to high-dimensional, geostatistical-based decision-support analyses}},
-url = {https://www.sciencedirect.com/science/article/pii/S1364815221000657},
-volume = {139},
-year = {2021}
-}
 @article{Madsen2003,
[…]
