+++ mikecalsetup/__init__.py
@@ -1,0 +1,3 @@
+from .main import Setup
+from .ost_file import OstFile
+from .ost_post_proc import OstPostProc
