+++ mikecalsetup/main.py
@@ -12,9 +12,10 @@ import os
 import subprocess
 import warnings
 import sys
-from mikecalsetup import ost_file
 import mikeio
 import pkg_resources
+from mikecalsetup import ost_file
+
 
 pd.options.mode.chained_assignment = "raise"
 
@@ -84,8 +85,9 @@ def extract_observations_from_file(stat_file, program='OSTRICH'):
     lines = f.readlines()
     f.close()
 
-    best_pos_values = {'me': 0, 'mea': 0, 'rms': 0, 'kge': 1, 'pcc': 1,
-                       'nse': 1, 'fbal': 0, 'ri': 0}
+    best_pos_values = {'me': 0, 'mae': 0, 'rmse': 0, 'kge': 1, 'pcc': 1,
+                       'nse': 1, 'fbal': 0, 'ri': 0, 'wbl': 0, 'kt': 1,
+                       'pbias': 0, 'apbias': 0, 'rsr': 0}
 
     # OSTRICH
     keys = [line.split(sep)[0] for line in lines]
@@ -365,8 +367,10 @@ def table_from_file(file):
                         'name', 'value', 'line', 'col']
     if len(parInidf) == 0:
         raise ValueError('Parameter df not read correctly or file empty')
+
     parInidf['value'] = parInidf['value'].str.replace("'", "", regex=True)
-    parInidf['value'] = parInidf['value'].str.replace("|", "", regex=True)
+    parInidf['value'] = parInidf['value'].str.replace("|", "", regex=False)
+    parInidf['value'] = parInidf['value'].str.replace(r'\|', '', regex=True)
 
     # clean up convert to floats, int and str.
     parInidf['value'] = parInidf['value'].apply(convert_str_to_appr_dtypes)
@@ -787,7 +791,7 @@ def pardf_to_pest_par(par_df):
     # Parameter groups section
     # * parameter groups
     # PARGPNME INCTYP DERINC DERINCLB FORCEN DERINCMUL DERMTHD
-    pargpnme = set([idx.split('_')[0] for idx in par_df.index])
+    pargpnme = list(set([idx.split('_')[0] for idx in par_df.index]))
     pest_parg = pd.DataFrame(index=pargpnme)
     pest_parg['inctyp'] = 'relative'
     pest_parg['derinc'] = 0.01
@@ -867,7 +871,6 @@ class ExtractParameters():
                   'TIME_SERIES_FILE': 'tsfile',
                   'ROOT_DEPTH': 'RD',
                   'GroundwaterTable': 'lr_GroundwaterTable'}
-
 
     def __init__(self, parInidf, proc_active, model_types, par_from, pth):
         self.parInidf = parInidf
@@ -1199,8 +1202,8 @@ class ExtractParameters():
         uz = self.parInidf.loc[self.parInidf['header0'] == 'Unsatzone'].copy()
 
         # settings
-        uz_bypass = int(uz.loc[uz.name == 'Bypass', 'value'])
-        uz_gaa = int(uz.loc[uz.name == 'USE_Green_And_Ampt', 'value'])
+        uz_bypass = uz.loc[uz.name == 'Bypass', 'value'].iloc[0]
+        uz_gaa = uz.loc[uz.name == 'USE_Green_And_Ampt', 'value'].iloc[0]
 
         # uztype = 0 (Richards), uztype = 1 (Gravity)
         if self.model_types['UZ'] in [0, 1]:
@@ -1282,7 +1285,7 @@ class ExtractParameters():
             sz = name_after(sz, 'Name')
 
             # settings
-            sz_wells = int(sz.loc[sz.name == 'Wells', 'value'])
+            sz_wells = sz.loc[sz.name == 'Wells', 'value'].iloc[0]
 
[…]
