### 2023-08-14 tenemark/mikecalsetup#1 @tenemark
Thank you again for taking the time to review this package! I greatly appreciate your feedback and insights, which will undoubtedly contribute to enhancing the user experience and reliability of this package. As suggested, I have created a conda environment file that includes a section for conda installations and another for pip installations.
