### 2023-08-14 tenemark/mikecalsetup#2 @tenemark
As suggested, I have included these in the environment file mike.yml.
