### 2023-08-14 tenemark/mikecalsetup#3 @tenemark
This specific path formatting issue that led to the OSError is due to the package not stripping the vertical lines from the string, when it should have. I have added an extra line to the script to avoid this error from happening again.
