### 2023-08-14 tenemark/mikecalsetup#4 @tenemark
The statement of need in the readme file have been modified to answer some of your questions. To answer your question about the observations, I have added a description of the processing to the example notebook “example_usage.ipynb”.
