### 2023-08-14 tenemark/mikecalsetup#5 @tenemark
As suggested, I have added some extra examples and moved them to the folder "examples". The examples now also include execution of a forward model run with PEST and OSTRICH. Also, an example (example_PEST_with_PstFrom.ipynb) shows how users of MIKE SHE can access the functionality of pyEMU and mikecalsetup at the same time.
