### 2023-08-14 tenemark/mikecalsetup#8 @tenemark
The package has a simple community guideline in the readme file.
