### 2023-08-28 tenemark/mikecalsetup#7 @tenemark
I have improved the test script and also included the example notebooks as part of the testing. Test coverage reporting is also an option now. I am still using unittest, but I will consider moving to pytest at a later date. The testing is not integrated yet either, but hopefully in the future I will find the time to set this up.
