+++ joss/paper.bib
@@ -2,7 +2,6 @@
 abstract = {A consistent framework for parameter estimation in distributed hydrological catchment modelling using automatic calibration is formulated. The framework focuses on the different steps in the estimation process from model parameterisation and selection of calibration parameters, formulation of calibration criteria,and choice of optimisation algorithm. The calibration problem is formulated in a general multi-objective context in which different objective functions that measure individual process descriptions can be optimised simultaneously. Within this framework it is possible to tailor the model calibration to the specific objectives of the model application being considered. A test example is presented that illustrates the use of the calibration framework for parameter estimation in the MIKE SHE integrated and distributed hydrological modelling system. A significant trade-off between the performance of the groundwater level simulations and the catchment runoff is observed in this case,defining a Pareto front with a very sharp structure. The Pareto optimum solution corresponding to a proposed balanced aggregated objective function is seen to provide a proper balance between the two objectives. Compared to a manual expert calibration,the balanced Pareto optimum solution provides generally better simulation of the runoff, whereas virtually similar performance is obtained for the groundwater level simulations. {\textcopyright} 2002 Elsevier Science Ltd. All rights reserved.},
 author = {Madsen, Henrik},
 doi = {10.1016/S0309-1708(02)00092-1},
-file = {:C\:/Users/wqg436/OneDrive - K{\o}benhavns Universitet/Papers/Madsen 2003.pdf:pdf},
 issn = {03091708},
 journal = {Advances in Water Resources},
 keywords = {Automatic calibration,Distributed model,Hydrological modelling,MIKE SHE,Multiple objectives,Parameter estimation},
@@ -22,7 +21,6 @@ year = {2017}
 }
 @article{Abbott1986,
 author = {Abbott, M.B. and Bathurst, J.C. and Cunge, J.A. and O'Connell, P.E. and Rasmussen, J.},
-file = {:C\:/Users/wqg436/OneDrive - K{\o}benhavns Universitet/Papers/Abott et al. 1986.pdf:pdf},
 issn = {00221694},
 journal = {Journal of Hydrology},
 pages = {61--77},
@@ -71,7 +69,6 @@ year = {2010}
 abstract = {We have developed pyEMU, a python framework for Environmental Modeling Uncertainty analyses, open-source tool that is non-intrusive, easy-to-use, computationally efficient, and scalable to highly-parameterized inverse problems. The framework implements several types of linear (first-order, second-moment (FOSM)) and non-linear uncertainty analyses. The FOSM-based analyses can also be completed prior to parameter estimation to help inform important modeling decisions, such as parameterization and objective function formulation. Complete workflows for several types of FOSM-based and non-linear analyses are documented in example notebooks implemented using Jupyter that are available in the online pyEMU repository. Example workflows include basic parameter and forecast analyses, data worth analyses, and error-variance analyses, as well as usage of parameter ensemble generation and management capabilities. These workflows document the necessary steps and provides insights into the results, with the goal of educating users not only in how to apply pyEMU, but also in the underlying theory of applied uncertainty quantification.},
 author = {White, Jeremy T. and Fienen, Michael N. and Doherty, John E.},
 doi = {10.1016/j.envsoft.2016.08.017},
-file = {:C\:/Users/wqg436/OneDrive - K{\o}benhavns Universitet/Papers/White et al 2016.pdf:pdf},
 issn = {13648152},
 journal = {Environmental Modelling and Software},
 keywords = {Data worth analysis,FOSM,Highly-parameterized,Model independent,Python,Uncertainty analysis},
