### 2023-08-29 #5548 @tenemark
Thank you again very much for your thorough review! Following your comments, I have made some revisions, which I will summarize here:
I have addressed the raised issues on github.
I have improved testing of the package with additional tests and coverage reporting.
I have added a conda environment file, which will ease installation of the package. By installing mikecalsetup in this environment, the test file will run successfully.
I have added some extra examples and moved them to the folder "examples". The examples now also include execution of a forward model run with PEST and OSTRICH. Also, an example (example_PEST_with_PstFrom.ipynb) shows how users of MIKE SHE can access the functionality of pyEMU and mikecalsetup at the same time.
I have changed the state of the field description to better convey how this package relate to pyEMU. The focus of mikecalsetup is to setup calibration files for the nested header text file format used in MIKE SHE files, which is not supported by pyEMU. Spatially distributed and time dependent data found in the dfs file format of MIKE SHE can be handled with pyEMU if they are converted to text files (which we now show in the example “example_PEST_with_PstFrom.ipynb”). mikecalsetup has therefore not been developed to handle these files.
I have edited the reference list and validated the DOIs.
