### 2023-09-12 tenemark/mikecalsetup#9 @jsta
Missing entry in bib file for JOSS
It looks to me like the bib file is missing an entry for `Andersson2023`

openjournals/joss-reviews#5548
