+++ joss/paper.bib
@@ -79,3 +79,11 @@ url = {http://dx.doi.org/10.1016/j.envsoft.2016.08.017},
 volume = {85},
 year = {2016}
 }
+@misc{Andersson2023,
+author = {Andersson, J. H. and Mariegaard, J. S.},
+title = {{MIKE IO}},
+year = {2023},
+publisher = {GitHub},
+journal = {GitHub repository},
+howpublished = {\url{https://github.com/DHI/mikeio}},
+}
