### 2023-09-15 5c30091
improve file path handling: more robust with os.path.join
