### 2023-09-15 5014cb0
add 'ofs' as attribute of class
