### 2023-09-15 c8a3208
plot_wsse(): do not plot OFs with value 0
