+++ mikecalsetup/ost_post_proc.py
@@ -167,14 +167,16 @@ class OstPostProc:
             for of in ofs:
                 if self.fs[of].max() == 0:
                     print(f'Warning: Not plotting {of} as all solutions have value 0')
-                    ofs.drop(of)
+                    ofs.remove(of)
         # Creating the scatter matrix pairplot
         if 'select' not in self.fs.columns:
             df = pd.concat([self.fs[ofs].assign(hue='All solutions'),
                             self.ns[ofs].assign(hue='Non-dom solutions')],
                             ignore_index=True)
             ax = sns.pairplot(df, hue='hue', diag_kind='kde',
-                              palette=['orange', 'k'])
+                              palette=['0.7', 'orange'], 
+                              markers=['X','o'])
+            ax.set(xlim=0, ylim=0)
         else:
             ss = self.fs.loc[self.fs.select == 1]
             df = pd.concat([self.fs[ofs].assign(hue='All solutions'),
@@ -182,7 +184,8 @@ class OstPostProc:
                             ss[ofs].assign(hue='Selected solutions')],
                             ignore_index=True)
             ax = sns.pairplot(df, hue='hue', diag_kind='kde',
-                              palette=['orange', 'k', 'blue'])
+                              palette=['0.7', 'orange', 'blue'], 
+                              markers=['X','o','o'])
         return ax
 
     def autoselect_solutions(self, method='pareto', ofs=None, of_weights=None,
