### 2023-09-15 69da12b
improve plot_wsse(): limit plots to 0, 95% quantiles
