### 2023-09-22 e1e85f7
read_ostoutput0(): Add development of pareto-front along optimization. Also read in class initialization
