### 2023-09-25 1db847b
get_observation_weights(): Now also return observatio names and df as float
