### 2023-09-25 45c75cd
autoselect_solution(): if reselect=False, give unique select markers to each new selection
