### 2023-09-25 45cbc3e
autoselect_solution(): 'weighted_ofs' uses range of pareto front solution OFs for scaling; simplify selection
