### 2023-10-05 #5548 @jsta
@tenemark The next step is for you to create a GitHub/Zenodo release:
https://github.com/tenemark/mikecalsetup/releases
