### 2023-10-06 #5548 @tenemark
Thank you! I have made a release.

On Thu, Oct 5, 2023 at 5:49 PM Jemma Stachelek ***@***.***>
wrote:

> @tenemark <https://github.com/tenemark> The next step is for you to
> create a GitHub/Zenodo release:
> https://github.com/tenemark/mikecalsetup/releases
>
