### 2023-10-09 #5548 @tenemark
I am unsure how to do this. I have changed the metadata of my Zenodo upload
to correspond to what you requested. The DOI is 10.5281/zenodo.8412800. If
this is not what you meant, could you please provide me with a link to a
guideline of what to do.

On Fri, Oct 6, 2023 at 2:51 PM Jemma Stachelek ***@***.***>
wrote:

> Great, next step @tenemark <https://github.com/tenemark> is making a DOI
> archive in Zenodo/Figshare/Etc. One critical piece of this step is that the
> archival deposit must have the correct metadata. This includes the title
> (should match the paper title) and author list (make sure the list is
> correct and people who only made a small fix are not on it). You may also
> add the authors' ORCID.
>
