### 2023-10-12 #5548 @kthyng
Congrats on your new publication @tenemark!! Many thanks to editor @jsta and reviewers @aleaf and @ecomodeller for your time, hard work, and expertise!!
